+++ lang/src/org/partiql/lang/eval/EvaluatingCompiler.kt
@@ -42,6 +42,7 @@ import org.partiql.lang.ast.sourceLocation
 import org.partiql.lang.ast.toAstStatement
 import org.partiql.lang.ast.toPartiQlMetaContainer
 import org.partiql.lang.domains.PartiqlAst
+import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.domains.staticType
 import org.partiql.lang.domains.toBindingCase
 import org.partiql.lang.errors.ErrorCode
@@ -1221,7 +1222,7 @@ internal class EvaluatingCompiler(
 
     private fun compileIs(expr: PartiqlAst.Expr.IsType, metas: MetaContainer): ThunkEnv {
         val expThunk = compileAstExpr(expr.value)
-        val typedOpParameter = expr.type.toTypedOpParameter(customTypedOpParameters)
+        val typedOpParameter = expr.type.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return thunkFactory.thunkEnv(metas) { valueFactory.newBoolean(true) }
         }
@@ -1265,7 +1266,7 @@ internal class EvaluatingCompiler(
 
     private fun compileCastHelper(value: PartiqlAst.Expr, asType: PartiqlAst.Type, metas: MetaContainer): ThunkEnv {
         val expThunk = compileAstExpr(value)
-        val typedOpParameter = asType.toTypedOpParameter(customTypedOpParameters)
+        val typedOpParameter = asType.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return expThunk
         }
@@ -1361,7 +1362,7 @@ internal class EvaluatingCompiler(
         thunkFactory.thunkEnv(metas, compileCastHelper(expr.value, expr.asType, metas))
 
     private fun compileCanCast(expr: PartiqlAst.Expr.CanCast, metas: MetaContainer): ThunkEnv {
-        val typedOpParameter = expr.asType.toTypedOpParameter(customTypedOpParameters)
+        val typedOpParameter = expr.asType.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return thunkFactory.thunkEnv(metas) { valueFactory.newBoolean(true) }
         }
@@ -1396,7 +1397,7 @@ internal class EvaluatingCompiler(
     }
 
     private fun compileCanLosslessCast(expr: PartiqlAst.Expr.CanLosslessCast, metas: MetaContainer): ThunkEnv {
-        val typedOpParameter = expr.asType.toTypedOpParameter(customTypedOpParameters)
+        val typedOpParameter = expr.asType.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return thunkFactory.thunkEnv(metas) { valueFactory.newBoolean(true) }
         }
@@ -3014,6 +3015,22 @@ internal class EvaluatingCompiler(
             },
             ordering
         )
+
+    /** Helper to convert [PartiqlAst.Type] in AST to a [TypedOpParameter]. */
+    private fun PartiqlAst.Type.toTypedOpParameter(): TypedOpParameter {
+        // hack: to avoid duplicating the function `PartiqlAst.Type.toTypedOpParameter`, we have to convert this
+        // PartiqlAst.Type to PartiqlPhysical.Type. The easiest way to do that without using a visitor transform
+        // (which is overkill and comes with some downsides for something this simple), is to transform to and from
+        // s-expressions again.  This will work without difficulty as long as PartiqlAst.Type remains unchanged in all
+        // permuted domains between PartiqlAst and PartiqlPhysical.
+
+        // This is really just a temporary measure, however, which must exist for as long as the type inferencer works only
+        // on PartiqlAst.  When it has been migrated to use PartiqlPhysical instead, there should no longer be a reason
+        // to keep this function around.
+        val sexp = this.toIonElement()
+        val physicalType = PartiqlPhysical.transform(sexp) as PartiqlPhysical.Type
+        return physicalType.toTypedOpParameter(customTypedOpParameters)
+    }
 }
 
 /**